Refactor ProjectTag to be generic
